requirements.txt
aiohttp==3.10.11
python-dotenv==1.1.1
python-socketio==5.13.0
